fix(npm): suggest flags after script name
